Extract container and related traits to a separate module
