+++ communication/src/message.rs
@@ -4,6 +4,7 @@ use std::sync::Arc;
 use bytes::arc::Bytes;
 use abomonation;
 use crate::{Data, Container};
+use crate::container::IntoAllocated;
 
 /// Either an immutable or mutable reference.
 pub enum RefOrMut<'a, T> where T: 'a {
@@ -257,261 +258,8 @@ impl<T: Container> Container for Message<T> {
     }
 }
 
-/// Turn a hollow allocation into an allocated object.
-pub trait IntoAllocated<T> {
-    /// Use the allocation in `self` to reconstruct the allocated object into an
-    /// owned object.
-    fn assemble(self, allocated: RefOrMut<T>) -> T where Self: Sized {
-        Self::assemble_new(allocated)
-    }
-    /// TODO
-    fn assemble_new(allocated: RefOrMut<T>) -> T;
-}
-
 impl<T: Container> IntoAllocated<Message<T>> for MessageAllocation<Option<T::Allocation>> {
     fn assemble_new(_allocated: RefOrMut<Message<T>>) -> Message<T> {
         unreachable!()
-    }
-}
-
-impl IntoAllocated<()> for () {
-    fn assemble_new(_allocated: RefOrMut<()>) -> () {
-        ()
-    }
-}
-
-impl<A: Container> IntoAllocated<(A, )> for (A::Allocation, )
-where
-    A::Allocation: IntoAllocated<A>,
-{
-    fn assemble(self, allocated: RefOrMut<(A, )>) -> (A, ) where Self: Sized {
-        match allocated {
-            RefOrMut::Ref(t) => (self.0.assemble(RefOrMut::Ref(&t.0)), ),
-            RefOrMut::Mut(t) => (self.0.assemble(RefOrMut::Mut(&mut t.0)), ),
-        }
-    }
-
-    fn assemble_new(allocated: RefOrMut<(A, )>) -> (A, ) {
-        match allocated {
-            RefOrMut::Ref(t) => (A::Allocation::assemble_new(RefOrMut::Ref(&t.0)), ),
-            RefOrMut::Mut(t) => (A::Allocation::assemble_new(RefOrMut::Mut(&mut t.0)), ),
-        }
-    }
-}
-
-impl<A: Container, B: Container> IntoAllocated<(A, B)> for (A::Allocation, B::Allocation)
-    where
-        A::Allocation: IntoAllocated<A>,
-        B::Allocation: IntoAllocated<B>,
-{
-    fn assemble(self, allocated: RefOrMut<(A, B)>) -> (A, B) where Self: Sized {
-        match allocated {
-            RefOrMut::Ref(t) => (
-                self.0.assemble(RefOrMut::Ref(&t.0)),
-                self.1.assemble(RefOrMut::Ref(&t.1))
-            ),
-            RefOrMut::Mut(t) => (
-                self.0.assemble(RefOrMut::Mut(&mut t.0)),
-                self.1.assemble(RefOrMut::Mut(&mut t.1))
-            ),
-        }
-    }
-
-    fn assemble_new(allocated: RefOrMut<(A, B)>) -> (A, B) {
-        match allocated {
-            RefOrMut::Ref(t) => (
-                A::Allocation::assemble_new(RefOrMut::Ref(&t.0)),
-                B::Allocation::assemble_new(RefOrMut::Ref(&t.1))
-            ),
-            RefOrMut::Mut(t) => (
-                A::Allocation::assemble_new(RefOrMut::Mut(&mut t.0)),
-                B::Allocation::assemble_new(RefOrMut::Mut(&mut t.1)),
-            ),
-        }
-    }
-}
-
-impl<A: Container, B: Container, C: Container> IntoAllocated<(A, B, C)> for (A::Allocation, B::Allocation, C::Allocation)
-    where
-        A::Allocation: IntoAllocated<A>,
-        B::Allocation: IntoAllocated<B>,
-        C::Allocation: IntoAllocated<C>,
-{
-    fn assemble(self, allocated: RefOrMut<(A, B, C)>) -> (A, B, C) where Self: Sized {
-        match allocated {
-            RefOrMut::Ref(t) => (
-                self.0.assemble(RefOrMut::Ref(&t.0)),
-                self.1.assemble(RefOrMut::Ref(&t.1)),
-                self.2.assemble(RefOrMut::Ref(&t.2)),
-            ),
-            RefOrMut::Mut(t) => (
-                self.0.assemble(RefOrMut::Mut(&mut t.0)),
-                self.1.assemble(RefOrMut::Mut(&mut t.1)),
-                self.2.assemble(RefOrMut::Mut(&mut t.2)),
-            ),
-        }
-    }
-
-    fn assemble_new(allocated: RefOrMut<(A, B, C)>) -> (A, B, C) {
-        match allocated {
-            RefOrMut::Ref(t) => (
-                A::Allocation::assemble_new(RefOrMut::Ref(&t.0)),
-                B::Allocation::assemble_new(RefOrMut::Ref(&t.1)),
-                C::Allocation::assemble_new(RefOrMut::Ref(&t.2)),
-            ),
-            RefOrMut::Mut(t) => (
-                A::Allocation::assemble_new(RefOrMut::Mut(&mut t.0)),
-                B::Allocation::assemble_new(RefOrMut::Mut(&mut t.1)),
-                C::Allocation::assemble_new(RefOrMut::Mut(&mut t.2)),
-            ),
-        }
-    }
-}
-
-impl<A: Container, B: Container, C: Container, D: Container> IntoAllocated<(A, B, C, D)> for (A::Allocation, B::Allocation, C::Allocation, D::Allocation)
-    where
-        A::Allocation: IntoAllocated<A>,
-        B::Allocation: IntoAllocated<B>,
-        C::Allocation: IntoAllocated<C>,
-        D::Allocation: IntoAllocated<D>,
-{
-    fn assemble(self, allocated: RefOrMut<(A, B, C, D)>) -> (A, B, C, D) where Self: Sized {
-        match allocated {
-            RefOrMut::Ref(t) => (
-                self.0.assemble(RefOrMut::Ref(&t.0)),
-                self.1.assemble(RefOrMut::Ref(&t.1)),
-                self.2.assemble(RefOrMut::Ref(&t.2)),
-                self.3.assemble(RefOrMut::Ref(&t.3)),
-            ),
-            RefOrMut::Mut(t) => (
-                self.0.assemble(RefOrMut::Mut(&mut t.0)),
-                self.1.assemble(RefOrMut::Mut(&mut t.1)),
-                self.2.assemble(RefOrMut::Mut(&mut t.2)),
-                self.3.assemble(RefOrMut::Mut(&mut t.3)),
-            ),
-        }
-    }
-
-    fn assemble_new(allocated: RefOrMut<(A, B, C, D)>) -> (A, B, C, D) {
-        match allocated {
-            RefOrMut::Ref(t) => (
-                A::Allocation::assemble_new(RefOrMut::Ref(&t.0)),
-                B::Allocation::assemble_new(RefOrMut::Ref(&t.1)),
-                C::Allocation::assemble_new(RefOrMut::Ref(&t.2)),
-                D::Allocation::assemble_new(RefOrMut::Ref(&t.3)),
-            ),
-            RefOrMut::Mut(t) => (
-                A::Allocation::assemble_new(RefOrMut::Mut(&mut t.0)),
-                B::Allocation::assemble_new(RefOrMut::Mut(&mut t.1)),
-                C::Allocation::assemble_new(RefOrMut::Mut(&mut t.2)),
-                D::Allocation::assemble_new(RefOrMut::Mut(&mut t.3)),
-            ),
-        }
-    }
-}
-
-impl<T: Clone> IntoAllocated<Vec<T>> for Vec<T> {
-    fn assemble(mut self, ref_or_mut: RefOrMut<Vec<T>>) -> Vec<T> {
-        match ref_or_mut {
-            RefOrMut::Ref(t) => {
-                self.clone_from(t);
-                println!("into_allocated<Vec<T>> ref")
-            }
-            RefOrMut::Mut(t) => {
-                ::std::mem::swap(&mut self, t);
-                println!("into_allocated<Vec<T>> mut")
-            },
-        }
-        self
-    }
-
-    fn assemble_new(ref_or_mut: RefOrMut<Vec<T>>) -> Vec<T> {
-        Self::new().assemble(ref_or_mut)
-    }
-}
-
-impl IntoAllocated<usize> for () {
-    fn assemble_new(allocated: RefOrMut<usize>) -> usize {
-        *allocated
-    }
-}
-
-impl IntoAllocated<String> for String {
-    fn assemble(mut self, allocated: RefOrMut<String>) -> String where Self: Sized {
-        match allocated {
-            RefOrMut::Ref(t) => self.clone_from(t),
-            RefOrMut::Mut(t) => ::std::mem::swap(&mut self, t),
-        }
-        self
-    }
-
-    fn assemble_new(allocated: RefOrMut<String>) -> String {
-        Self::with_capacity(allocated.len()).assemble(allocated)
-    }
-}
-
-mod rc {
-    use crate::message::{IntoAllocated, RefOrMut};
-    use std::rc::Rc;
-
-    impl<T> IntoAllocated<Rc<T>> for () {
-        fn assemble(self, ref_or_mut: RefOrMut<Rc<T>>) -> Rc<T> {
-            Rc::clone(&*ref_or_mut)
-        }
-
-        fn assemble_new(ref_or_mut: RefOrMut<Rc<T>>) -> Rc<T> {
-            Rc::clone(&*ref_or_mut)
-        }
-    }
-}
-
-#[cfg(test)]
-mod test {
-    use crate::Container;
-    use crate::message::{IntoAllocated, RefOrMut};
-
-    #[test]
-    fn test_container_ref() {
-        let c = vec![1, 2, 3, 4];
-        let data = c.hollow().assemble(RefOrMut::Ref(&vec![5, 6, 7, 8]));
-        assert_eq!(data, vec![5, 6, 7, 8]);
-    }
-
-    #[test]
-    fn test_container_mut() {
-        let c = vec![1, 2, 3, 4];
-        let data = c.hollow().assemble(RefOrMut::Mut(&mut vec![5, 6, 7, 8]));
-        assert_eq!(data, vec![5, 6, 7, 8]);
-    }
-
-    #[test]
-    fn test_tuple_len() {
-        let t = (1, vec![1, 2, 3], "Hello".to_string());
-        assert_eq!(t.len(), 9);
-    }
-
-    #[test]
-    fn test_vec_hollow() {
-        let c = vec![1, 2, 3, 4].hollow();
-        assert!(c.is_empty());
-        assert_eq!(c.capacity(), 4);
-    }
-
-    #[test]
-    fn test_vec_assemble_ref() {
-        let c = vec![1, 2, 3, 4].hollow();
-        let c = c.assemble(RefOrMut::Ref(&vec![5, 6]));
-        assert!(!c.is_empty());
-        assert_eq!(c.capacity(), 4);
-        assert_eq!(c.len(), 2);
-    }
-
-    #[test]
-    fn test_vec_assemble_mut() {
-        let c = vec![1, 2, 3, 4].hollow();
-        let c = c.assemble(RefOrMut::Mut(&mut vec![5, 6]));
-        assert!(!c.is_empty());
-        assert_eq!(c.capacity(), 2);
-        assert_eq!(c.len(), 2);
     }
 }
